+++ apps/ranger-next/components/Header/header.tsx
@@ -1,0 +1,9 @@
+const Header = () => {
+  return (
+    <header className="header">
+      <p>Header Nav</p>
+    </header>
+  )
+}
+
+export default Header